+++ CHANGELOG.md
@@ -86,3 +86,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ## [Unreleased] - 2022-04-17
 ### Added
 - Added new plotting script subclone_relative_seed_variability.R
+
+## [Unreleased] - 2022-04-23
+### Added
+- Added new plotting script plot_num_subclones_box.R